reuse duplicated code
